support WebSocket frame fragmentation
Hi!
I'm working with a websocket server which sends messages containing a lot of data. It's doing it by using websocket's feature called `Fragmentation` according to RFC6455 https://tools.ietf.org/html/rfc6455#section-5.4.

The problem is that currently only the first `WebSocketFrame` passed to `WebsocketHandler` which leads to corrupted messages.

To support this feature, websocket client should be able to handle fragmented messages.
Netty already has an aggregator precisely for this task, so I propose just to use it.